feat: new left day card added
