+++ lib/config/theme/app_theme.dart
@@ -3,6 +3,6 @@ import 'package:flutter/material.dart';
 abstract class AppTheme {
   static ThemeData getTheme() => ThemeData(
     useMaterial3: false,
-    brightness: Brightness.dark
+    brightness: Brightness.dark,
   );
 }
