--- lib/utils/get_week_name.dart
@@ -1,12 +1,0 @@
-String getWeekName(int day) {
-  final List<String> daysOfWeek = <String>[
-		'Monday',
-		'Tuesday',
-		'Wednesday',
-		'Thursday',
-		'Friday',
-		'Saturday',
-		'Sunday'
-	];
-  return daysOfWeek[day - 1];
-}
